# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface AccessStatus {
  authorized: boolean;
  isAdmin: boolean;
  adminConfigured: boolean;
}

export interface InvitationInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  userId: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  label: string;
}

export interface Invitation {
  userId: string;
  label: string;
  createdAt: string;
}

export interface SourceFile {
  name: string;
  size: number;
}

export type DashboardStatusSource = typeof DashboardStatusSource[keyof typeof DashboardStatusSource];


export const DashboardStatusSource = {
  uploaded_snapshot: 'uploaded_snapshot',
  sharepoint: 'sharepoint',
} as const;

export interface DashboardStatus {
  source: DashboardStatusSource;
  snapshotDate: string;
  quarter: string;
  /** @nullable */
  lastAttempt: string | null;
  /** @nullable */
  lastSuccess: string | null;
  /** @nullable */
  error: string | null;
  validation: string;
  warnings: string[];
  files: SourceFile[];
  version: string;
  refreshing: boolean;
}

export interface HealthStatus {
  status: string;
}

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};