# lib/api-zod/src/generated/types/accessStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccessStatus {
  authorized: boolean;
  isAdmin: boolean;
  adminConfigured: boolean;
}