# lib/api-zod/src/generated/types/dashboardStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatusSource } from './dashboardStatusSource';
import type { SourceFile } from './sourceFile';

export interface DashboardStatus {
  source: DashboardStatusSource;
  snapshotDate: string;
  quarter: string;
  /** @nullable */
  lastAttempt: string | null;
  /** @nullable */
  lastSuccess: string | null;
  /** @nullable */
  error: string | null;
  validation: string;
  warnings: string[];
  files: SourceFile[];
  version: string;
  refreshing: boolean;
}