# lib/api-zod/src/generated/types/invitation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Invitation {
  userId: string;
  label: string;
  createdAt: string;
}