# lib/api-zod/src/generated/types/invitationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InvitationInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  userId: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  label: string;
}